Add files via upload
This is the script containing all steps. 
I would have separated the main components of the analysis, for example into data exploration, PCA & upsampling, and then model training and testing. 
But with the jupyter notebook formats, we would have to save the data and load it again  for each respective script. Which would simply take too long. 
I am looking into implementing separate scripts for the final repository.