Add files via upload
This is the final corrected script with the updates.
The main changes are implemented to account for trouble with the computational load of the previous script. 
A new section has been added reducing the features in our dataset by binning them into ranges and obtaining the mean from those bins.
The width of the bin can also easily be altered to be larger and therefore further decrease the number of features.